refactor to vm
